requirements.txt
asgiref==3.8.1
Django==5.1.5
django-extensions==3.2.3
djangorestframework==3.15.2
pillow==10.4.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
